requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.3
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mod_wsgi==5.0.2
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
soupsieve==2.6
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.21.0